+++ src/terrains/staircase.h
@@ -8,6 +8,8 @@ typedef struct Staircase Staircase;
 #include <stdlib.h>
 #include <stdbool.h>
 #include "../character_listings.h"
+#include "../dungeon.h"
+#include "../events.h"
 
 struct Staircase {
     u_char x;
@@ -22,5 +24,7 @@ struct Staircase {
 
 Staircase* staircase_initialize(u_char x, u_char y, u_char fromFloor, u_char toFloor);
 Staircase* staircase_terminate(Staircase* staircase);
+
+Dungeon* staircase_take(Dungoen* dungeon, Staircase* staircase);
 
 #endif